+++ bootstrap_modal_forms/forms.py
@@ -2,9 +2,9 @@ from django import forms
 from bootstrap_modal_forms.mixins import PopRequestMixin, CreateUpdateAjaxMixin
 
 
-class BSForm(PopRequestMixin, forms.Form):
+class BSModalForm(PopRequestMixin, forms.Form):
     pass
 
 
-class BSModalForm(PopRequestMixin, CreateUpdateAjaxMixin, forms.ModelForm):
+class BSModalModelForm(PopRequestMixin, CreateUpdateAjaxMixin, forms.ModelForm):
     pass